Adiciona busca paginada de jogos utilizando Elasticsearch e ajusta testes e dependências

+++ GamePlatform.Jogos.Application/DTOs/ResultadoPaginadoDto.cs
@@ -5,5 +5,5 @@ public class ResultadoPaginadoDto<T>
     public IEnumerable<T> Itens { get; init; } = [];
     public int NumeroPagina { get; init; }
     public int TamanhoPagina { get; init; }
-    public int TotalDeItens { get; init; }
+    public long TotalDeItens { get; init; }
 }
